pyproject.toml
[project]
name = "llmfinetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.2.0",
    "einops>=0.8.1",
    "huggingface-hub>=0.28.1",
    "numpy>=2.2.2",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "peft>=0.14.0",
    "python-dotenv>=1.0.1",
    "scipy>=1.15.2",
    "sentencepiece>=0.2.0",
    "torch>=2.6.0",
    "torchtext>=0.18.0",
    "tqdm>=4.67.1",
    "transformers>=4.48.3",
    "trl>=0.14.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "pyright>=1.1.393",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
]
